refactor: replace albedoIndex -1 sentinel with std::optional
primitive_data::albedoIndex was an int defaulting to -1 for "no texture",
forcing the backend to bounds-check and cast at the use site. Same pattern
uniform locations were moved to std::optional<int> for. Now
std::optional<std::size_t>: presence replaces the sentinel, the cast is
gone, and the "no texture" case is a plain emptiness test.

Generated-By: pi 0.85.1 (https://pi.dev)

+++ cgame-graphics/source/gltf_loader.cpp
@@ -163,7 +163,7 @@ namespace cgame::graphics
                         const auto& gltfTexture = asset.textures[textureIndex];
 
                         if (gltfTexture.imageIndex)
-                            data.albedoIndex = static_cast<int>(*gltfTexture.imageIndex);
+                            data.albedoIndex = *gltfTexture.imageIndex;
                     }
                 }
